Refactor(#169): 테스트 코드 작성을 위한 컴포넌트 수정
- 하나의 테스트에서 mocking이 많아지는 것을 방지하지 위하여 컴포넌트를 분리리
